Allow the gem to be used from a read-only location
Some environments install their gem files and then make everything read-only.
This gem fails when installed in such environments because it tries to open its
own version file in 'r+' mode, just to read the version from it.
This commit allows the read options to be passed in for anyone who wants to use
the gem just for reading a version file without modifying it.